+++ src/main/java/de/jeff_media/angelchest/hooks/EcoEnchantsHook.java
@@ -1,5 +1,6 @@
 package de.jeff_media.angelchest.hooks;
 
+import com.jeff_media.jefflib.PDCUtils;
 import de.jeff_media.angelchest.Main;
 import de.jeff_media.angelchest.config.Config;
 import org.bukkit.NamespacedKey;
@@ -9,6 +10,8 @@ import org.bukkit.event.entity.PlayerDeathEvent;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.Objects;
+
 /**
  * This is for situations when a player gets killed by another player with the Telekinesis enchantment.
  */
@@ -16,6 +19,7 @@ public class EcoEnchantsHook {
 
     private static final Enchantment telekinesisEnchant;
     private static final Main main = Main.getInstance();
+    private static final NamespacedKey soulboundKey = Objects.requireNonNull(NamespacedKey.fromString("ecoenchants:soulbound"));
 
     static {
         telekinesisEnchant = getTelekinesisEnchant();
@@ -35,6 +39,18 @@ public class EcoEnchantsHook {
     @Nullable
     private static Enchantment getTelekinesisEnchant() {
         return Enchantment.getByKey(NamespacedKey.minecraft("telekinesis"));
+    }
+
+    public static boolean isEcoEnchantsSoulbound(ItemStack item) {
+        if(item == null) return false;
+        if(!item.hasItemMeta()) return false;
+        boolean result = PDCUtils.has(item, soulboundKey);
+        if(result) {
+            if(main.debug) {
+                main.debug("Item is EcoEnchants soulbound: " + item);
+            }
+        }
+        return result;
     }
 
 }